Cleanup

+++ src/image_1d_example.rs
@@ -12,7 +12,7 @@ pub struct Args {
     /// WARNING, if this Directory
     /// already exists, current contents will be removed.
     #[arg(short, long)]
-    pub output_dir: std::path::PathBuf,
+    pub output_dir: Option<std::path::PathBuf>,
 
     /// Chunk size to use for parallelism.
     #[arg(short, long, default_value = "1000")]
@@ -79,12 +79,14 @@ impl Args {
             build_info::print_report(name);
         }
 
-        let output_dir = args.output_dir.to_str().unwrap();
-        let _ = std::fs::remove_dir_all(output_dir);
-        std::fs::create_dir(output_dir).unwrap();
+        if let Some(output_dir) = &args.output_dir {
+            let output_dir = output_dir.to_str().unwrap();
+            let _ = std::fs::remove_dir_all(output_dir);
+            std::fs::create_dir(output_dir).unwrap();
+        }
 
-        let mut output_image_path = args.output_dir.clone();
-        output_image_path.push(format!("{}.png", name));
+        let mut output_image_path = args.output_dir.as_ref().unwrap().clone();
+        output_image_path.push("image.png");
 
         rayon::ThreadPoolBuilder::new()
             .num_threads(args.threads)
@@ -104,6 +106,12 @@ impl Args {
 
     pub fn grid_bounds(&self) -> AABB<1> {
         AABB::new(matrix![0, self.domain_size as i32 - 1])
+    }
+
+    pub fn dot_path(&self) -> PathBuf {
+        let mut dot_path = self.output_dir.as_ref().unwrap().clone();
+        dot_path.push("plan.dot");
+        dot_path
     }
 
     pub fn save_wisdom(&self) {
