+++ src/image_2d_example.rs
@@ -12,7 +12,7 @@ pub struct Args {
     /// WARNING, if this Directory
     /// already exists, current contents will be removed.
     #[arg(short, long)]
-    pub output_dir: std::path::PathBuf,
+    pub output_dir: Option<std::path::PathBuf>,
 
     /// Chunk size to use for parallelism.
     #[arg(short, long, default_value = "1000")]
@@ -79,9 +79,11 @@ impl Args {
             build_info::print_report(name);
         }
 
-        let output_dir = args.output_dir.to_str().unwrap();
-        let _ = std::fs::remove_dir_all(output_dir);
-        std::fs::create_dir(output_dir).unwrap();
+        if let Some(output_dir) = &args.output_dir {
+            let output_dir = output_dir.to_str().unwrap();
+            let _ = std::fs::remove_dir_all(output_dir);
+            std::fs::create_dir(output_dir).unwrap();
+        }
 
         rayon::ThreadPoolBuilder::new()
             .num_threads(args.threads)
@@ -106,9 +108,15 @@ impl Args {
     }
 
     pub fn frame_name(&self, i: usize) -> PathBuf {
-        let mut result = self.output_dir.to_path_buf();
+        let mut result = self.output_dir.as_ref().unwrap().clone();
         result.push(format!("frame_{:04}.png", i));
         result
+    }
+
+    pub fn dot_path(&self) -> PathBuf {
+        let mut dot_path = self.output_dir.as_ref().unwrap().clone();
+        dot_path.push("plan.dot");
+        dot_path
     }
 
     pub fn save_wisdom(&self) {
